the Job type button is added now :D
